📈added app insights tracking for root layout & page specific exceptions
